Move more things and fix tests
