+++ kube-core/Cargo.toml
@@ -11,10 +11,23 @@ license = "Apache-2.0"
 repository = "https://github.com/clux/kube-rs"
 readme = "../README.md"
 
+[features]
+ws = []
+
 [dependencies]
 serde = { version = "1.0.118", features = ["derive"] }
+serde_json = "1.0.61"
+thiserror = "1.0.23"
+once_cell = "1.7.2"
+url = "2.2.0"
+log = "0.4.11"
+http = "0.2.2"
 
 [dependencies.k8s-openapi]
 version = "0.11.0"
 default-features = false
 features = []
+
+[dev-dependencies]
+kube = { path = "../kube", version = "^0.53.0" }
+
