+++ kube-runtime/Cargo.toml
@@ -35,6 +35,7 @@ rustls-tls = ["kube/rustls-tls"]
 
 [dev-dependencies]
 kube-derive = { path = "../kube-derive", version = "^0.53.0"}
+kube-core = { path = "../kube-core", version = "^0.53.0"}
 serde_json = "1.0.61"
 tokio = { version = "1.0.1", features = ["full", "test-util"] }
 rand = "0.8.0"
