+++ kube/Cargo.toml
@@ -21,7 +21,7 @@ native-tls = ["openssl", "hyper-tls", "tokio-native-tls"]
 rustls-tls = ["hyper-rustls", "tokio-rustls"]
 derive = ["kube-derive"]
 jsonpatch = ["json-patch"]
-ws = ["tokio-tungstenite", "rand"]
+ws = ["tokio-tungstenite", "rand", "kube-core/ws"]
 oauth = ["tame-oauth"]
 gzip = ["async-compression"]
 admission = ["json-patch"]
@@ -67,7 +67,6 @@ tame-oauth = { version = "0.4.7", features = ["gcp"], optional = true }
 pin-project = "1.0.4"
 rand = { version = "0.8.3", optional = true }
 tracing = "0.1.25"
-once_cell = "1.7.2"
 
 [dependencies.k8s-openapi]
 version = "0.11.0"
